requirements.txt
langgraph>=0.0.20
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
pymongo>=4.5.0
tiktoken>=0.5.0